+++ PharmacyMicroservice/src/main/java/com/medex/services/PharmacyService.java
@@ -4,7 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.medex.communicationmodules.PharmacyInfo;
+import com.medex.communicationmodules.Status;
 import com.medex.database.PharmacyDB;
+
 import com.medex.model.PharmaceuticalStock;
 import com.medex.model.Pharmacy;
 
@@ -22,40 +24,44 @@ public class PharmacyService {
 		List<PharmacyInfo> pharmacyinfoList = new ArrayList<PharmacyInfo>(); //Make a list that contains HostInfo instances
 		if (pharmacyList.isEmpty() == false) return null;	
 		//If the list of hosts is not empty, "cast" every host instance into a HostInfo instance
-		int i = 0; //Could have just used a for loop rather than a foreach loop with a counter really..
-		for (Pharmacy p : pharmacyList) { //For each host, "cast" it into a HostInfo instance and add it it to the hostinfoList
-			pharmacyinfoList.add(new PharmacyInfo(p));
-			List<PharmaceuticalStock> lst = pharmaceuticalStockService.getPharmaceuticalStocks(p.getId()); //For every host, get the list of VMs it holds and attach that to the hashmap of VMs that each instance of HostInfo has.
-			if (lst.isEmpty() == false) { //If not empty (There are VMs attached)
-				pharmacyinfoList.get(i++).listToMap(lst); //Copy the list of VMs.
-				}
+		for (Pharmacy p : pharmacyList) {
+			pharmacyinfoList.add(PharmacyToPharmacyInfo(p));
 			}		
 		return pharmacyinfoList;
 	}
 	
-	public PharmacyInfo getPharmacy(int id)
+	public PharmacyInfo getPharmacy(int pharmacyid)
 	{
-		Pharmacy pharmacy = pharmacydb.getPharmacy(id); //Get all hosts.
+		Pharmacy pharmacy = pharmacydb.getPharmacy(pharmacyid); //Get all hosts.
 		if (pharmacy == null) return null;
-		PharmacyInfo pharmacyInfo = new PharmacyInfo(pharmacy);
-		List<PharmaceuticalStock> lst = pharmaceuticalStockService.getPharmaceuticalStocks(pharmacyInfo.getId()); //For every host, get the list of VMs it holds and attach that to the hashmap of VMs that each instance of HostInfo has.
-		if (lst.isEmpty() == false) { pharmacyInfo.listToMap(lst); 	}	
-		return pharmacyInfo;
+		return PharmacyToPharmacyInfo(pharmacy);
 	}
 	
-	public Pharmacy addPharmacy(Pharmacy aPharmacy)
+	public PharmacyInfo addPharmacy(Pharmacy aPharmacy)
 	{
-		pharmacydb.insertPharmacy(aPharmacy); return aPharmacy;
+		pharmacydb.insertPharmacy(aPharmacy); return PharmacyToPharmacyInfo(aPharmacy);
 	}
 	
-	public Pharmacy updatePharmacy(Pharmacy aPharmacy)
+	public PharmacyInfo updatePharmacy(Pharmacy aPharmacy)
 	{
-		pharmacydb.updatePharmacy(aPharmacy); return aPharmacy;
+		pharmacydb.updatePharmacy(aPharmacy); return PharmacyToPharmacyInfo(aPharmacy);
 	}
 	
-	public void removePharmacy(int id)
+	public Status removePharmacy(int id)
 	{
+		if (pharmacydb.getPharmacy(id) == null) return new Status(false);
 		pharmacydb.deletePharmacy(id);
+		return new Status(true);
+	}
+	
+	
+	private PharmacyInfo PharmacyToPharmacyInfo(Pharmacy aPharmacy)
+	{
+		PharmacyInfo pharmacyInfo = new PharmacyInfo(aPharmacy);
+		List<PharmaceuticalStock> lst = pharmaceuticalStockService.getPharmaceuticalStocks(aPharmacy.getId()); //For every host, get the list of VMs it holds and attach that to the hashmap of VMs that each instance of HostInfo has.
+		if (lst.isEmpty() == false) { pharmacyInfo.listToMap(lst); }
+		
+		return pharmacyInfo;
 	}
 }
